+++ CHANGELOG.md
@@ -25,6 +25,7 @@
   - Add group size support for convolutional layers  
   - Add quantized matmul support for for convolutional layers  
   - Add 7-bit, 5-bit and 3-bit quantization support  
+  - Add separate quant mode option for Text Encoders  
   - Fix forced FP32 with tensorwise FP8 matmul  
   - Fix PyTorch <= 2.4 compatibility with FP8 matmul  
   - Fix VAE with conv quant